implement proper loop over all outputs, implement utility functions needed for that loop, refactoring

+++ core/qmckcpp/include/qmck/generic_table_format.hpp
@@ -7,11 +7,11 @@ namespace qmck
     struct generic_table_format
     {
     public:
-        std::uint32_t inputs_count{};
-        std::uint32_t outputs_count{};
+        std::uint8_t inputs_count{};
+        std::uint8_t outputs_count{};
 
     public:
-        generic_table_format(uint32_t inputs_count, uint32_t outputs_count);
+        generic_table_format(uint8_t inputs_count, uint8_t outputs_count);
 
         generic_table_format() = default;
